chore: separate RowScript from EditorScropt
