+++ frontend/rust-lib/flowy-grid/tests/grid/block_test/mod.rs
@@ -1,0 +1,5 @@
+#![allow(clippy::module_inception)]
+mod block_test;
+mod row_test;
+mod script;
+pub mod util;
