+++ frontend/rust-lib/flowy-grid/tests/grid/cell_test/mod.rs
@@ -1,0 +1,2 @@
+mod script;
+mod test;
