+++ frontend/rust-lib/flowy-grid/tests/grid/field_test/test.rs
@@ -1,6 +1,6 @@
-use crate::grid::field_util::*;
-use crate::grid::script::EditorScript::*;
-use crate::grid::script::*;
+use crate::grid::field_test::script::FieldScript::*;
+use crate::grid::field_test::script::GridFieldTest;
+use crate::grid::field_test::util::*;
 use flowy_grid::services::field::select_option::SelectOption;
 use flowy_grid::services::field::SingleSelectTypeOption;
 use flowy_grid_data_model::revision::TypeOptionDataEntry;
@@ -8,23 +8,23 @@ use flowy_sync::entities::grid::FieldChangesetParams;
 
 #[tokio::test]
 async fn grid_create_field() {
-    let mut test = GridEditorTest::new().await;
-    let (params, field_rev) = create_text_field(&test.grid_id);
+    let mut test = GridFieldTest::new().await;
+    let (params, field_rev) = create_text_field(&test.grid_id());
 
     let scripts = vec![
         CreateField { params },
         AssertFieldEqual {
-            field_index: test.field_count,
+            field_index: test.field_count(),
             field_rev,
         },
     ];
     test.run_scripts(scripts).await;
 
-    let (params, field_rev) = create_single_select_field(&test.grid_id);
+    let (params, field_rev) = create_single_select_field(&test.grid_id());
     let scripts = vec![
         CreateField { params },
         AssertFieldEqual {
-            field_index: test.field_count,
+            field_index: test.field_count(),
             field_rev,
         },
     ];
@@ -33,9 +33,9 @@ async fn grid_create_field() {
 
 #[tokio::test]
 async fn grid_create_duplicate_field() {
-    let mut test = GridEditorTest::new().await;
-    let (params, _) = create_text_field(&test.grid_id);
-    let field_count = test.field_count;
+    let mut test = GridFieldTest::new().await;
+    let (params, _) = create_text_field(&test.grid_id());
+    let field_count = test.field_count();
     let expected_field_count = field_count + 1;
     let scripts = vec![
         CreateField { params: params.clone() },
@@ -47,11 +47,11 @@ async fn grid_create_duplicate_field() {
 
 #[tokio::test]
 async fn grid_update_field_with_empty_change() {
-    let mut test = GridEditorTest::new().await;
-    let (params, field_rev) = create_single_select_field(&test.grid_id);
+    let mut test = GridFieldTest::new().await;
+    let (params, field_rev) = create_single_select_field(&test.grid_id());
     let changeset = FieldChangesetParams {
         field_id: field_rev.id.clone(),
-        grid_id: test.grid_id.clone(),
+        grid_id: test.grid_id(),
         ..Default::default()
     };
 
@@ -59,7 +59,7 @@ async fn grid_update_field_with_empty_change() {
         CreateField { params },
         UpdateField { changeset },
         AssertFieldEqual {
-            field_index: test.field_count,
+            field_index: test.field_count(),
             field_rev,
         },
     ];
@@ -68,14 +68,14 @@ async fn grid_update_field_with_empty_change() {
 
 #[tokio::test]
 async fn grid_update_field() {
-    let mut test = GridEditorTest::new().await;
-    let (params, single_select_field) = create_single_select_field(&test.grid_id);
+    let mut test = GridFieldTest::new().await;
+    let (params, single_select_field) = create_single_select_field(&test.grid_id());
 
     let mut single_select_type_option = SingleSelectTypeOption::from(&single_select_field);
     single_select_type_option.options.push(SelectOption::new("Unknown"));
     let changeset = FieldChangesetParams {
         field_id: single_select_field.id.clone(),
-        grid_id: test.grid_id.clone(),
+        grid_id: test.grid_id(),
         frozen: Some(true),
         width: Some(1000),
         type_option_data: Some(single_select_type_option.protobuf_bytes().to_vec()),
@@ -92,7 +92,7 @@ async fn grid_update_field() {
         CreateField { params },
         UpdateField { changeset },
         AssertFieldEqual {
-            field_index: test.field_count,
+            field_index: test.field_count(),
             field_rev: expected_field_rev,
         },
     ];
@@ -101,9 +101,9 @@ async fn grid_update_field() {
 
 #[tokio::test]
 async fn grid_delete_field() {
-    let mut test = GridEditorTest::new().await;
-    let original_field_count = test.field_count;
-    let (params, text_field_rev) = create_text_field(&test.grid_id);
+    let mut test = GridFieldTest::new().await;
+    let original_field_count = test.field_count();
+    let (params, text_field_rev) = create_text_field(&test.grid_id());
     let scripts = vec![
         CreateField { params },
         DeleteField {
