+++ frontend/rust-lib/flowy-grid/tests/grid/mod.rs
@@ -1,8 +1,5 @@
 mod block_test;
 mod cell_test;
 mod field_test;
-mod field_util;
 mod filter_test;
-mod row_test;
-mod row_util;
-mod script;
+mod grid_editor;
